requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
aiofiles
pydantic